01.09 login change 2nd

+++ config/functions.php
@@ -8,21 +8,17 @@
 		$action = $_GET['action'];
 	} else {
 		echo "<script>alert('Illegal Operation !');window.location.href='../admin/index.php';</script>";
+		exit;
 	}
 	
 	switch($action){
 		//登录判断
 		case 'login':
-			$email = $_POST['email'];
-			$password = md5($_POST['password']);
-
-			$pattern = "/^([0-9A-Za-z\\-_\\.]+)@([0-9a-z]+\\.[a-z]{2,3}(\\.[a-z]{2})?)$/i";
-			if ( preg_match( $pattern, $email ) ) {
-				continue;
-			} else {
-				echo "<script>alert('Illegal E-mail Address !');window.location.href='../admin/index.php';</script>";
-				exit;
+			$email = filter_var(($_POST['email']),FILTER_VALIDATE_EMAIL);
+			if (!$email) {
+			    throw new InvalidArgumentException('invalid email');
 			}
+			$password = md5($_POST['password']);
 
 			//拼接select语句执行得到结果，并跳转
 			$sql = "SELECT * from `user` where `email`='{$email}' and `password`='{$password}'";
@@ -30,6 +26,7 @@
 			if ($res && mysqli_num_rows($res)>0) {
 				$user = mysqli_fetch_assoc($res);
 				$_SESSION['user']=$user['id'];	//存session
+				session_regenerate_id();
 				echo "<script>alert('验证成功');window.location.href='../admin/index.php';</script>";
 			} else {
 				echo "<script>alert('验证失败');window.location.href='../admin/login.php';</script>";
@@ -40,9 +37,9 @@
 		//添加文章
 		case 'add':
 			loginCheck();
-			$title = $_POST['title'];
+			$title = mysqli_real_escape_string($_POST['title']);
 			$formaltext = mysqli_real_escape_string($link,$_POST['formaltext']);//拼接sql专用的字符串处理
-			$column = $_POST['column'];
+			$column = mysqli_real_escape_string($_POST['column']);
 
 			//拼接insert语句,执行，得到结果
 			$sql = "INSERT into `article`(title,formaltext,`column`) VALUES ('{$title}','{$formaltext}','{$column}');";
@@ -58,15 +55,15 @@
 		//修改文章
 		case 'edit':
 			loginCheck();
-			$id = $_GET['id'];
-			$title = $_POST['title'];
+			$id = filter_var(($_GET['id']),FILTER_VALIDATE_INT);
+			$title = mysqli_real_escape_string($_POST['title']);
 			$formaltext = mysqli_real_escape_string($link,$_POST['formaltext']);	//拼接sql专用的字符串处理
-			$column = $_POST['column'];
+			$column = mysqli_real_escape_string($_POST['column']);
 
 			//拼接update语句,执行，判断并跳转
 			$sql = "UPDATE `article` set `title`='{$title}',`formaltext`='{$formaltext}',`column`='{$column}' where `id`={$id};";
 			$res = mysqli_query($link,$sql);
-			if($res){
+			if ($res) {
 				echo "<script>alert('修改成功');window.location.href='../admin/index.php';</script>";
 			} else {
 				echo "<script>alert('修改失败');window.location.href='../admin/edit.php?id={$id}';</script>";
@@ -77,12 +74,15 @@
 		//删除文章
 		case 'delete':
 			loginCheck();
-			$id = $_GET['id'];
+			$id = filter_var(($_GET['id']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
+			if (!$id) {
+			    throw new InvalidArgumentException('invalid id');
+			}
 			
 			//拼接delete语句并跳转
 			$sql = "DELETE from `article` where `id`={$id}";
 			$res = mysqli_query($link,$sql);
-			if($res){
+			if ($res) {
 				echo "<script>alert('删除成功');window.location.href='../admin/index.php';</script>";
 			} else {
 				echo "<script>alert('删除失败');window.location.href='../admin/index.php';</script>";
@@ -93,8 +93,8 @@
 		//添加标签
 		case 'tag':
 			loginCheck();
-			$text = $_POST['text'];
-			$id = $_POST['id'];
+			$text = mysqli_real_escape_string($_POST['text']);
+			$id = filter_var(($_POST['id']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
 			
 			//执行，判断并跳转
 			$sql = "UPDATE `article` set `tag`='{$text}' where `id`={$id};";
@@ -106,8 +106,8 @@
 		//添加标签
 		case 'addTag':
 			loginCheck();
-			$articleId = $_POST['articleId'];
-			$tagId = $_POST['id'];
+			$articleId = filter_var(($_POST['articleId']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
+			$tagId = filter_var(($_POST['id']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
 
 			//
 			$sql = "SELECT * from article where `id`='{$articleId}'";
@@ -126,8 +126,8 @@
 		//删除标签
 		case 'reduceTag':
 			loginCheck();
-			$articleId = $_POST['articleId'];
-			$tagId = $_POST['id'];
+			$articleId = filter_var(($_POST['articleId']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
+			$tagId =filter_var(($_POST['id']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
 
 			//查找该文章的所有标签
 			$sql = "SELECT * from article where `id`='{$articleId}'";
@@ -156,7 +156,7 @@
 		case 'newTag':
 			loginCheck();
 			$text = $_POST['text'];
-			$articleId= $_POST['id'];
+			$articleId = filter_var(($_POST['id']),FILTER_VALIDATE_INT,array('options' => array('min_range' => 1)));
 
 			//拼接insert语句,执行，得到结果
 			$sql = "INSERT into `tag` (`name`) VALUES ('{$text}');";
@@ -186,9 +186,8 @@
 
 	function loginCheck() 
 	{
-		if(empty($_SESSION['user'])){
+		if(!isset($_SESSION['user'])){
 			echo "<script>alert('请登录');window.location.href='../admin/login.php';</script>";
 			exit;
 		}
-		session_regenerate_id();
 	}
